Eliminar y Modificar Administradores

+++ src/UI/DetalleUsuario.java
@@ -5,6 +5,7 @@
  */
 package UI;
 
+import BL.BL_Logueo;
 import BL.BL_ManejadorUsuario;
 import BL.BL_Usuario;
 import java.awt.Cursor;
@@ -17,6 +18,7 @@ import javax.swing.JOptionPane;
  */
 public class DetalleUsuario extends javax.swing.JFrame {
     BL_Usuario usuario;
+    BL_Logueo sesion;
     BL_ManejadorUsuario manejador;
     ArrayList<BL_Usuario> listaTotalUsuarios;
     int identificador;
@@ -256,6 +258,10 @@ public class DetalleUsuario extends javax.swing.JFrame {
 
     private void checkAdministradorActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_checkAdministradorActionPerformed
         // TODO add your handling code here:
+        if(!manejador.BuscarAdministradores() && usuario.isAdministrador()){
+            JOptionPane.showMessageDialog(null, "Error","Error",JOptionPane.ERROR_MESSAGE);
+            checkAdministrador.setSelected(true);
+        }
     }//GEN-LAST:event_checkAdministradorActionPerformed
 
     private void btnPanelModificarMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnPanelModificarMouseClicked
@@ -297,18 +303,34 @@ public class DetalleUsuario extends javax.swing.JFrame {
         if(txtNombreCompleto.getText().trim().equals("") || txtNombreUsuario.getText().trim().equals("") || txtContrasena.getText().trim().equals("")){
             JOptionPane.showMessageDialog(null, "Los campos deben estar completos","Error",JOptionPane.ERROR_MESSAGE);
         }else{            
-        
+            
             usuario.setIdUsuario(listaTotalUsuarios.get(identificador).getIdUsuario());
             if(JOptionPane.showConfirmDialog(null, "¿Desea eleminar este usuario?", "Eliminar usuario", JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION){
-                if(usuario.eliminarUsuario()){
-                    JOptionPane.showMessageDialog(null, "El usuario seleccionado se eliminó correctamente del sistema","Eliminación exitosa",JOptionPane.INFORMATION_MESSAGE);
-                    manejador.Eliminar(identificador);
-
-                    listaTotalUsuarios = manejador.ObtenerListaUsuarios();
-                    this.dispose();
-                    new AdministrarUsuarios().setVisible(true);
+                if(usuario.getIdUsuario() != sesion.getIdUsuario()){
+                    if(usuario.eliminarUsuario()){
+                        manejador.Eliminar(identificador);
+                        JOptionPane.showMessageDialog(null, "El usuario seleccionado se eliminó correctamente del sistema","Eliminación exitosa",JOptionPane.INFORMATION_MESSAGE);
+                        
+                        listaTotalUsuarios = manejador.ObtenerListaUsuarios();
+                        this.dispose();
+                        new AdministrarUsuarios().setVisible(true);
+                    }
                 }else{
-                    JOptionPane.showMessageDialog(null, "Error al eliminar el usuario seleccionado","Error",JOptionPane.ERROR_MESSAGE);
+                    if(manejador.BuscarAdministradores()){
+                        if(JOptionPane.showConfirmDialog(null, "¿Desea eliminar su cuenta?\n" + "Se cerrará la sesión y no podra ingresar con sus credenciales en otra ocasión", "Confirmar eliminación", JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION){
+                            if(usuario.eliminarUsuario()){
+                                manejador.Eliminar(identificador); 
+                                JOptionPane.showMessageDialog(null, "El usuario seleccionado se eliminó correctamente del sistema","Eliminación exitosa",JOptionPane.INFORMATION_MESSAGE);
+                                
+                                listaTotalUsuarios = manejador.ObtenerListaUsuarios();
+                                        
+                                this.dispose();
+                                new Login().setVisible(true);
+                            }
+                        }
+                    }else{
+                        JOptionPane.showMessageDialog(null, "El usuario que desea eliminar es un Administrador y no existen más administradores en el sistema","Error",JOptionPane.ERROR_MESSAGE);
+                    }
                 }
             }
         }
